Split forms to preventing render loop
When production builds, Vue doesn't not detecting infinite loops.
and 'value' with fuction(not in `data()`) cause infinite loops.

The 'value' may has multiple values
So I split selector for targeting per-value.

#43
