+++ src/components/OptionsModule.vue
@@ -4,25 +4,7 @@
     <div class="module-desc">{{meta.description}}</div>
     <div class="module-options">
       <div class="module-option" v-for="option in meta.options">
-        <!-- DROPDOWN -->
-        <div v-if="option.type === 'dropdown'">
-          {{option.title}}
-          <select :data-module="meta.name" :data-optid="option.id" :data-notice="option.notice" :value="config.get(option.id)" v-on:change="dropdownChange">
-            <option :value="choice.value" v-for="choice in option.data">{{choice.name}}</option>
-          </select>
-        </div>
-        <!-- RADIO -->
-        <div v-else-if="option.type === 'radio'">
-
-        </div>
-        <!-- CHECK -->
-        <div v-else-if="option.type === 'check'">
-          {{option.title}} <input type="checkbox" :data-module="meta.name" :data-optid="option.id" :data-notice="option.notice" v-on:change="checkChange" :checked="config.get(option.id) ? 'checked' : ''">
-        </div>
-        <!-- TEXT -->
-        <div v-else>
-          {{option.title}} <input type="text" :data-module="meta.name" :data-optid="option.id" :data-notice="option.notice" :value="config.get(option.id)" v-on:change="textChange">
-        </div>
+        <OptionsSelector :meta="meta" :option="option" :config="config" :notice="notice" />
       </div>
     </div>
     <div class="module-notice" v-show="notice">{{notice}}</div>
@@ -31,6 +13,7 @@
 
 <script>
 import HioriSDK from '@sdk'
+import OptionsSelector from './OptionsSelect.vue'
 export default {
   props: [ 'meta' ],
   data() {
@@ -39,24 +22,8 @@ export default {
       notice: ''
     }
   },
-  methods: {
-    dropdownChange: function(evt){
-      this.updateConfig(evt.target.getAttribute('data-optid'), evt.target.value, evt.target.getAttribute('data-notice'))
-    },
-    textChange: function(evt){
-      this.updateConfig(evt.target.getAttribute('data-optid'), evt.target.value, evt.target.getAttribute('data-notice'))
-    },
-    checkChange: function(evt){
-      this.updateConfig(evt.target.getAttribute('data-optid'), evt.target.checked, evt.target.getAttribute('data-notice'))
-    },
-    updateConfig: function(optId, value, notice){
-      if (notice) this.notice = notice
-      this.config.set(optId, value)
-    }
-  },
-  mounted() {
-    // this.config = new HioriSDK.Options(this.meta.name)
-    this.$forceUpdate()
+  components: {
+    OptionsSelector
   }
 }
 </script>
